+++ src/js/components/internal/slider-preload.js
@@ -1,37 +1,9 @@
-import { $$, isVisible, removeAttr } from 'uikit-util';
+import Lazyload from '../../mixin/lazyload';
 
 export default {
+    mixins: [Lazyload],
+
     connected() {
-        if (window.IntersectionObserver) {
-            this.observer = new IntersectionObserver(
-                (entries) => {
-                    if (entries.some((entry) => entry.isIntersecting)) {
-                        removeLazyLoad(this.getAdjacentSlides());
-                    }
-                },
-                { rootMargin: '50% 50%' }
-            );
-            this.observer.observe(this.$el);
-        }
-    },
-
-    disconnected() {
-        this.observer && this.observer.disconnect();
-    },
-
-    update: {
-        read() {
-            if (isVisible(this.$el)) {
-                removeLazyLoad(this.getAdjacentSlides());
-            }
-        },
-
-        events: ['resize'],
+        this.lazyload(this.slides, this.getAdjacentSlides);
     },
 };
-
-export function removeLazyLoad(elements = []) {
-    for (const el of elements) {
-        el && $$('img[loading="lazy"]', el).forEach((el) => removeAttr(el, 'loading'));
-    }
-}