+++ src/js/core/switcher.js
@@ -1,3 +1,5 @@
+import Lazyload from '../mixin/lazyload';
+import Swipe from '../mixin/swipe';
 import Togglable from '../mixin/togglable';
 import {
     $$,
@@ -11,13 +13,14 @@ import {
     hasClass,
     matches,
     queryAll,
+    ready,
     toggleClass,
     toNodes,
     within,
 } from 'uikit-util';
 
 export default {
-    mixins: [Togglable],
+    mixins: [Lazyload, Swipe, Togglable],
 
     args: 'connect',
 
@@ -26,7 +29,6 @@ export default {
         toggle: String,
         itemNav: String,
         active: Number,
-        swiping: Boolean,
     },
 
     data: {
@@ -34,7 +36,6 @@ export default {
         toggle: '> * > :first-child',
         itemNav: false,
         active: 0,
-        swiping: true,
         cls: 'uk-active',
         attrItem: 'uk-switcher-item',
     },
@@ -79,6 +80,17 @@ export default {
                 this.toggles.some((toggle) => within(toggle, child))
             );
         },
+
+        swipeTarget() {
+            return this.connects;
+        },
+    },
+
+    connected() {
+        this.lazyload(this.$el, this.connects);
+
+        // check for connects
+        ready(() => this.$emit());
     },
 
     events: [
@@ -136,27 +148,21 @@ export default {
 
         show(item) {
             const prev = this.index();
-            const next = getIndex(
-                this.children[getIndex(item, this.toggles, prev)],
-                children(this.$el)
-            );
-
-            if (prev === next) {
-                return;
-            }
-
-            this.children.forEach((child, i) => {
-                toggleClass(child, this.cls, next === i);
-                attr(this.toggles[i], 'aria-expanded', next === i);
+            const next = getIndex(item, this.toggles, prev);
+            const active = getIndex(this.children[next], children(this.$el));
+            children(this.$el).forEach((child, i) => {
+                toggleClass(child, this.cls, active === i);
+                attr(this.toggles[i], 'aria-expanded', active === i);
             });
 
+            const animate = prev >= 0 && prev !== next;
             this.connects.forEach(async ({ children }) => {
                 await this.toggleElement(
                     toNodes(children).filter((child) => hasClass(child, this.cls)),
                     false,
-                    prev >= 0
+                    animate
                 );
-                await this.toggleElement(children[next], true, prev >= 0);
+                await this.toggleElement(children[active], true, animate);
             });
         },
     },